src/requirements.txt
# Core Dependencies - Pinned for App Service Stability
python-dotenv==1.0.1

# HTTP Server and Client - Resolved Dependency Conflicts
aiohttp==3.11.11
yarl==1.18.3
multidict==6.1.0
async-timeout==5.0.1

# Microsoft Teams Bot Framework - Enhanced SDK 4.17
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botbuilder-integration-aiohttp==4.17.0

# Teams AI Library - Production Ready
teams-ai>=1.6.0,<2.0.0

# Azure Services - GA Versions
azure-storage-blob==12.24.0
azure-ai-agents==1.0.2
azure-identity==1.23.0
azure-core==1.34.0
azure-search-documents==11.5.3

# AI/ML Dependencies - Semantic Kernel & SK Dependencies  
semantic-kernel==1.14.0
openai==1.58.1
numpy==2.2.1
pydantic==2.10.4

# Production Server
gunicorn==23.0.0

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Additional Production Dependencies
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# Crypto and Security
cryptography==45.0.2

# Date/Time Utilities
isodate==0.7.2
typing-extensions==4.13.2